commenting and cleaning stdioplusplus header

+++ headers/stdioplusplus.h
@@ -3,21 +3,19 @@
 
 #include <stdio.h>
 
-// returns the size of a file
-int fsize(char *file_path);
+// returns the size of a null-terminated file.
+size_t fsize(char *file_path);
 
-// returns all bytes of a file in a secure way (all buffers are allocated using sodium)
+// returns all bytes comprised between the beginning of the file and EOF in a secure way.
 unsigned char *fgetalls(char *file_path);
 
-// the same as above but for not sensitive data (all buffers are allocated using standard malloc)
+// the same as above but for not sensitive data.
 unsigned char *fgetall(char *file_path);
 
-// this function gets the file content from start_pos (inclusive)
-// to the end of the stream. It uses only sodium-allocated buffers.
+// returns all bytes comprised between start_pos (inclusive) and EOF in a secure way.
 unsigned char *fgetfroms(char *file_path, int start_pos);
 
-// this function gets the file content from start_pos (inclusive)
-// to end_pos (exclusive). It uses only sodium-allocated buffers.
+// returns all bytes comprised between start_pos (inclusive) and end_pos (exclusive) in a secure way.
 unsigned char *fgetfromtos(char *file_path, int start_pos, int end_pos);
 
 #endif
